Rework the logic for extracting the count value from a string
The work I recently did to add count-checking functionality to the validation
script has meant that I've looked a lot more at the output of the parsing
function, and this has revealed a bug.

The impetus behind this change was thus to fix that bug, which was caused by not
interpreting all whitespace as valid separators in a counting string. That's
solved by the call "".join(x.split()), which gets rid of all whitespace in the
string.

Once that was done, I realised that the regex logic can be considerably
simplified, since we only ever look at the first matching substring.