Update the url-finding routine to search for new short links as well
Reddit has introduced a short link format which looks like
`reddit.com/r/subreddit/s/{gibberish}`. In order to extract a (submission_id,
comment_id) pair from that we need to make a request to reddit with the short
link, and then check what url we get redirected to.

This commit adds that functionality to the find urls in text routine