🐛: Use all alphanumeric characters as possible values in ORD
The horribly complicated code for only repeating digits can be given a parameter
"n", which tells the class what the universe of digits to consider is. In
principle, we can do this for all bases, but only the base 10 path is ever used.

However, since the functionality is there, we might as well do all the bases
2-36 properly