Add Isenary side thread
Describing what a count in isenary looks like is fairly easy; according to the
rather loose criteria used elsewhere in the project, it's just a string which
contains one of the tokens ["they're", "taking", "the", "hobbits", "to",
"Isengard"].

This is just a modified base 6, so we could be satisfied with saying that all
threads are 1296  counts long. Unfortunately, r/counting decided a while ago to
shorten/lengthen some of the threads to make sure that the phrase "they're
taking the hobbits to isengard" was a get. Understandable, but we're still
recovering from that, and the threads currently have varying length.

So, I've implemented code which will calculate the actual count based on the
text in a comment. To do this it:

1. Takes a mapping of tokens to values [or a list of tokens, with implicit
mapping given by their index]
2. Takes the first line of the text, removes some irrelevant characters and
splits the line into words
3. Does fuzzy matching for each word against each token to see what the best
match is
4. Returns the integer formed by interpreting the concatenation of the value of
each token as a base n integer

This obviously will work for other threads, not just this one, but that's for
next time