Split up enormous side thread file
At more than 800 lines it was getting too big, and I'd known that for a while,
but the issue was finding a sensible way to split it up.

I'm not sure this is the best way of doing things, but I've tried to separate
the code which creates the framework for how I deal with side threads from the
code which uses that framework.

That gives the following spit:

- `rules.py`, which contains the classes that can check whether a count is valid
time- or wait-wise
- `validate_form.py` which contains functions to answer the question of whether
a given comment looks like a count under various rules
- `validate_count.py` which contains functions to help calculate how far along a
sequence of counts a given comment might be
- `side_threads.py` which contains code that composes the above three pieces of
information into a coherent whole.
- `thread_list.py` which is basically a huge list of side thread objects, one
for each known thread.