Add periodic table side thread
Getting to this point needs a bit of preparation, so this is a larger commit
than some of the others

The `elements.txt` file just contains a list of all the chemical symbols, and
is not particularly interesting

In principle, once we have the value of each element, getting the value of a
concatenation of elements can be done by using the `count from word list`
function.

Unfortunately, things aren't quite that simple. The function previously assumed
that words/tokens would be separated by whitespace, and this assumption was
hardcoded into the function.

Chemical formulas generally aren't separated like that, and neither are counts.
So step one of this commit was to split out the tokenization functionality and
make `count from word list` take a tokenizer as a parameter, with the default
tokenizer reproducing the previous behaviour.

Once that was done, I just had to make a tokenizer for chemical formulas. For
now, nobody's thought of adding numbers to them, so we just need to split before
every capital letter, which can be accomplished by using the regex
"[A-Z][^A-Z]*".