Add US states side thread
This needs a few changes to the existing code, as well as some extra data to be
added.

There's an alphabetical list of US states which is loaded in, and used in the
-`from character list`functions.

However, the default tokenizxer has assumed that tokens did not contain
whitespace, so it split the input string on whitespace and compared each word in
the string with the list of tokens.

A different approach is needed for this one, so that function has been modified.
We now also split each token on whitespace before comparing. If the token we
identify as the most likely candidate for a given word is non-final (i.e. part
of a multi-word token and not at the end), we grab the next word of the input
string and use that as well on the full tokens.

This approach works for one and two word tokens; for more complicated cases it
falls apart. If i ever get to there, I should probably see if I can find a
library that does this for me.