Add mostly repeating digits and no consecutive digits
... By doing a big-ass refactoring

Mostly repeating digits is very similar to only repeating digits, so it makes
sense to pull out the shared logic into a base class which they both inherit
from. Once that's done, adding no consecutive digits is simplish, but means that
the base class should change a little bit to accomodate a different number of
valid states.

A DFA-like side thread needs two (well 3) pieces of information to work:

- A description of which states are valid words in the side thread
- A way of calculating all words of length k. This is to speed up calculations,
  and could in principle be ignored

On top of those things, it also needs to know how many states there are for a
digit in a word. For example, ORD tracks whether a digit has been seen 0, 1 or 2
or more times, while NCD only tracks whether a digit has been seen or not.