Add no consecutive side thread
This is the first one where I couldn't be bothered to try and work out a
way of calculating the number of complete words of length k.

And that required quite a few changes to the rest of the code, since otherwise
pylint complains.

So the trick is to remove the NotImplemented complete_words from the base class
and replace it by a noop, and then override that implementation in the
classes that actually do something.

To make everything work, the result is wrapped in a tuple that says whether
anything was done, which lets the base class finish the call correctly