Refactor the dfa side threads
Precalculating the result for some threads has been a fun mathematical exercise,
but it's not too easy to follow, and the code feels brittle.

It was sort of necessary as a speedup for the ORD and MRD threads, since their
state space was of size 3**k, and calculating the powers of the (3**n, 3**n)
transition matrix took an appreciable amount of time..

After the work of adding the shorter state description for the homogeneous
threads, the state for those threads was (k + 2)C2, which is so much smaller
that there are no problems wwwwwwwith time any more.

That meant that I could remove the `complete words` function from all the
DFASideThread classes, and simplify the logic in DFASideThread.count, which no
longer had to check a flag of whether any precalculation had been done.

After that was done, the only functionality really left in all the subclasses of
DFASideThread was the calculation of which indices are considered a success.
But that's not really a stateful calculation, so we might as well calculate it
in an ordinary function and just pass it to the init of DFASideThread. And once
all of that is done there's no real reason to keep the concrete classes as
subclasses of DFASideThread, when they're just plain instances of the class.