🐛: Use latest submission when notifying about already-logged threads
Previously, the script would accidentally just print the name of the thread the
logging script started from, even when it had gone further back in time. That's
because when the script detected that the submission was already logged, the
comment variable is never updated. The fix is to use the submission variable
directly, rather than going through the comment variable.