Remove unneeded method override
Looking at the source code of `models.py`, it seems that the `parent` method of
CommentTree is identical to that in Tree, with the different behaviour of the
two arising from the fact that CommentTree has overridden the `node` method. But
that means we can get rid of the `parent` override, since the correct `node`
method is called anyway.