Manually add exponential backoff to the thread logging script
Praw unfortunately sometimes gets 429 errors from overuse of the API - I don't
quite understand why, since I thought they had built-in code to respect the api
limits.

Anyway, when we get a `TooManyRequests` exception, we should wait a bit and then
try again. The exact values probably don't matter, but we'll wait for 30s first
time, and then an extra factor of 1.5 times for each time logging the same
thread fails.

Unfortunately, adding this check made pylint unhappy, so we've had to update the
limit in the configuration file. I did try looking at the function to see if
there's an easy way of splitting it to have fewer branches, but I don't think
there is.