Use the NAOT caching to greatly simplify the class
If we're going to have a state->int dictionary and an int->state dictionary
and use those to work with our states, then there's no need for the ints to
actually be meaningful.

We can just loop over all our states, discarding the invalid ones, and assign an
increasing integer to each valid state we encounter.

That lets us get rid of all the state to int and int to state logic, and allows
us to replace function calls with dictionary lookups.

As a bonus, this results in a slight speedup.