Bail when encountering incorrect links in side thread stats
We have no robust way of mapping a side thread submission to its get. So when
people accidentally link the submission itself instead of the relevant comment,
getting that to work can be a bit tricky.

Previously, the code would just crash in that situation; now we just pretend
that we're done with that side thread and move on to the next one.