Handle empty FTF posts
The side thread stats post checks the current FTF post to see if the stats
comment has already been posted. Previously, this check raised an error if the
post had no comments. This commit fixes that, and ensures that in the case where
there are no comments, the check returns "no duplicate found".