Add `offset` parameter to errors function
Sometimes, there's a break in a thread and some counts happen on a side chain
rather than on the main thread. That means they'll be out of sync with what they
should be according to their position in the linear chain. To still be able to
detect errors in this shifted chain, an offset parameter is used, which tells us
how much the chain is shifted with respect to the true chain.

It allows for better detection of errors in these chains, with less noise.