Update how links are written in the directory
Previously, reddit would interpret a link of the form [link
test](/submission_id) as a markdown link to that reddit submission. They've
changed their backend code, and all the links are broken.

Luckily the submission ids are still valid, and replacing the above with [link
text](/comments/submission_id) yields a link to the correct place.

This commit updates the parsing code to handle both forms of the link in parsing
directory rows, and updates the row class to use the new style links when
writing.

Luckily, we don't use these original links very much, so nobody has complained
about the broken links yet.