Parse directory pages line by line instead of paragraph by paragraph
See [here](https://old.reddit.com/r/counting/comments/1ffs2dc/free_talk_friday_472/lnuzby0/)

The logic of the directory parser has always been to ignore as much of the
markdown as possible. The updater only knows about tables -- and very specific
tables at that -- so the parser has been set up to identify these tables and
leave the rest of the text alone.

It used to do this by splitting the text into paragraphs on empty lines, and
asking whether all the lines in a paragraph looked like table entries. If they
did, the whole paragraph was marked as table, and updated in the directory
updater. If just one of the lines didn't look like a table, the whole paragraph
was marked as text, and ignored in all further processing.

In markdown, lines starting with "#" are automatically considered headers, even
if there is no blank line before them, which would make the above parsing give
an incorrect result.

This commit fixes that, by going line by line rather than paragraph by
paragraph, and for each line asking whether it looks like a table or like text.
It builds up a buffer of "table" lines or "text" lines, and when the type
switches, it calls that a paragraph, writes it to the output variable, and
flushes the buffer.

This means that the parsed text is delivered in the same format as before, and
no changes to the consuming code are needed.