Automatically update the top 25 table in the directory
Previously, this table was sorted, but rows were not added or removed from it
automatically.

The exception was when a top 25 thread got archived.

The reason for this was that promoting and demoting rows would require removing
the promoted rows from the other tables, and adding the demoted rows back
somewhere sensible.

Removing the promoted rows turned out to be easier than feared (knock on wood),
and the demoted rows can just be added to the miscellaneous table at the bottom
of the directory.