Add levenshtein dependency
I was getting annoyed with the "slow-python sequence matcher" and the project
seems stable, so I added it as a dependency. I haven't done it before because it
felt excessive, and I considered making it an optional dependency, but I think
that would need a bit more work. The right way of doing it would probably be to
create a "side threads" dependency group and all all the things that only
involve parsing, validating and "to-count"-ing side threads to that group. Off
the top of my head, fuzzy matching and scipy would be included there, and if
there are any things that are only used by the finite state machines they should
probably go there too.