Handle mixed old-style and new-style links when seeking urls
Previously, new-style urls would always be considered after old-style urls,
which might be incorrect. A better approach is to return the urls in the order
they appear in the submission self text.

To do that we need to tag each match for either regular expression with the
position of the start of the match, do all the processing we need to do on the
matches themselves, and finally sort the processed urls by the match position.