Remove historical correction in "is_even" function
Some threads accidentally start at the wrong place. For the most part (and
specifically for the threads involved here) this has been corrected in the
database itself, so that ad hoc corrections like this one aren't needed any
more. That has the great advantage of being clearer to understand and fixing
pandas warnings at the same time ;)