Check the title of the previous submission before picking a link
Sometimes people do silly things and link submissions which are definitely not
the previous ones in the thread. One way we can check this is by looking at the
title of a purported previous post.

This commit implements the following algorithm.

- Take the first complete link (submission_id & comment_id are present) with a
title similarity score greater than some threshold
- If there are none, take the last link where the title has a similarity score
greater than the threshold
- If there are none, take the first link in the post
- If there are no links, just return (none, none)

Writing this out made me realise that there's a potential bug when all the links
present are to the same post., which for some reason has a completely different
title, and where the first link is just to a post and the next is to a comment.

Hm