Change depth finding function to be iterative rather than recursive
The default max recursion limit is 1000 and counting threads can get so long
that we actually hit that limit, which is annoying. The only thing we were
actually using the stack for was to keep track of what the current depth was, so
instead of doing that, we can start from the root (where we know the depth),
and go through the children of every node, with children having a depth of one
more than the parents. That way, there's absolutely no recursion in sight