Add "--archive" option to the side thread updater
This lets us optionally update the database of side threads based on threads
that are currently archived. Thse threads can be very old -- from before the
time when r/c had a rough consensus on what a side thread should look like --
so currently, only threads that are known to the directory updater will be
tracked. Adding a thread to the directory updater isn't very hard, so doing it
this way ensures that we only track threads that the directory updater would
actually understand