Don't follow links that lead to http 403 errors
Look, people sometimes link really weird things in the body of reddit
submissions. Sometimes, they link directly to people's user pages, and sometimes
these users get banned, causing 403 errors when trying to access the links. That
makes the directory updater crash, and me sad. The solution is to catch that
error and just continue, hoping that there's an actual link somewhere else in
the post or in the top level comments