Add an exponential backoff sleep routine when walking up threads
I don't know what happened, but praw is getting quite bad at backing off by
istelf, or reddit is getting quite bad at telling it it needs to back off. In
any case, I'm getting an increasing number of http 429 errors, so this is a
bandaid to avoid some of them.