Add a column to the database tracking the leaf submission
We do this for each thread in the directory, so that we can tell at a glance
whether the current submission in the directory is the same as it used to be. If
it is, then there are no new submissions to log, so there's no point in wasting
API requests on finding that out.