Refactor side thread clases
The structure of the class was slightly odd, in a way which became very much
apparent after I wrote a Base N side thread subclass. The rules element and the
count form/valid count/comment to count functionality is completely orthogonal,
so perhaps it's better to move that to separate classes and then just use the
side thread class to compose the two subclasses.

I'm not sure I've found the perfect solution yet -- the side thread lists have
become a lot more noisy. On the other hand, the new approach lets me validate
more threads, including all the special rules threads, which were previously
checked for invalid counts, but not for incorrect counts